+++ src/const.ml
@@ -44,7 +44,9 @@ let home = match (Sys.getenv "HOME") with
   | None -> failwith "Wrong value for home\n"
 ;;
 
-(* Some settings variables *)
-let rc_file = home ^ "/" ^ ".oclaunch_rc.json";;
+(* Default place to read settings *)
+let rc_file_default = home ^ "/" ^ ".oclaunch_rc.json";;
+(* Current place to read settings, maybe modified from command line argument *)
+let rc_file = ref rc_file_default;;
 (* Set tmp file, in witch stock launches, in biniou format *)
 let tmp_file = "/tmp/.oclaunch_trace.dat";; (* File where launch are logged *)