+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Changelog of OcLaunch
 
 ## v0.2.1
- +
+ + Add new command line option:
+   + “-c file” allow to read configuration from custom file.
 
 ## v0.2.0
  + Add new command line option.